Added a simple text file to record last download
The file sits in the same location as the downloaded files. It is a
text file where the first line is a date time string specifying when
the last dataset was downloaded. The methods added are
get_local_metadata and set_local_metadata.

The prompt was also changed to give information to the user.